src/requirements.txt
flask
google-auth
google-auth-httplib2
google-api-python-client
google-cloud-firestore
python-dotenv
